+++ src/Column/ColumnFactory.php
@@ -10,6 +10,8 @@ use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Schema\Column\AbstractColumnFactory;
 use Yiisoft\Db\Schema\Column\ColumnInterface;
 
+use Yiisoft\Db\Syntax\ColumnDefinitionParserInterface;
+
 use function hex2bin;
 use function str_starts_with;
 use function substr;
@@ -116,5 +118,10 @@ final class ColumnFactory extends AbstractColumnFactory
         }
 
         return parent::normalizeNotNullDefaultValue($defaultValue, $column);
+    }
+
+    protected function columnDefinitionParser(): ColumnDefinitionParserInterface
+    {
+        return new ColumnDefinitionParser();
     }
 }